UNG-1643 - extracting common code
